[Bronze IV] Title: Bank Transfer, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/21633. Bank Transfer/Bank Transfer.cc
@@ -3,6 +3,6 @@
 int main() {
   int k;
   scanf("%d", &k);
-  double c = 25 + k / 100.;
-  printf("%.2lf", c < 100 ? 100 : c > 2000 ? 2000 : c);
+  float c = 25 + k / 100.;
+  printf("%.2f", c < 100 ? 100 : c > 2000 ? 2000 : c);
 }
